# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Exalent Episode API",
    "version": "1.0.0",
    "description": "API documentation for Exalent Episode operations"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/updatePassword/{email}/{newPassword}": {
      "put": {
        "summary": "Update a user password",
        "tags": [
          "Password"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "email",
            "required": true,
            "description": "enter users email to update password",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "newPassword",
            "required": true,
            "description": "enter new password",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Password and metadata updated successfully",
            "content": {
              "application/json": {
                "example": {
                  "message": "Password and metadata updated successfully"
                }
              }
            }
          },
          "400": {
            "description": "Bad request, invalid input",
            "content": {
              "application/json": {
                "example": {
                  "message": "Invalid input"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "example": {
                  "message": "User not found"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "example": {
                  "message": "Internal server error"
                }
              }
            }
          }
        }
      }
    },
    "/resetPassword/{email}": {
      "put": {
        "summary": "Reset email for a user to change password",
        "tags": [
          "Password"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "email",
            "required": true,
            "description": "enter users email",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Password and metadata updated successfully",
            "content": {
              "application/json": {
                "example": {
                  "message": "Password and metadata updated successfully"
                }
              }
            }
          },
          "400": {
            "description": "Bad request, invalid input",
            "content": {
              "application/json": {
                "example": {
                  "message": "Invalid input"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "example": {
                  "message": "User not found"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "example": {
                  "message": "Internal server error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}